% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_level.R
\name{data_action}
\alias{data_action}
\alias{data_status}
\alias{lock_data}
\alias{unlock_data}
\title{Perform data-level locking operations}
\usage{
data_action(
  record,
  event = NULL,
  instrument = NULL,
  instance = NULL,
  arm = NULL,
  action,
  redcap_uri,
  token
)

data_status(
  record,
  event = NULL,
  instrument = NULL,
  instance = NULL,
  arm = NULL,
  redcap_uri,
  token
)

lock_data(
  record,
  event = NULL,
  instrument = NULL,
  instance = NULL,
  arm = NULL,
  redcap_uri,
  token
)

unlock_data(
  record,
  event = NULL,
  instrument = NULL,
  instance = NULL,
  arm = NULL,
  redcap_uri,
  token
)
}
\arguments{
\item{record}{A vector of REDCap record names to affect. Required.}

\item{event}{A character vector specifying the event/s to affect. Ignored if REDCap project is not longitudinal}

\item{instrument}{A character vector specifying the instrument/s to affect. If NULL, all instruments for the given record/event specification are affected}

\item{instance}{A vector specifying which instance of a repeating instrument to affect. If not specified, affects every instance.}

\item{arm}{A vector specifying which arm to affect.}

\item{action}{A character string for the action to perform on records. One of "lock", "unlock" or "status"}

\item{redcap_uri}{A character string of the REDCap server's API endpoint URL}

\item{token}{A character string of the REDCap API token}
}
\description{
Perform data-level locking operations
}
\section{Functions}{
\itemize{
\item \code{data_status}: Get the lock status of specified instruments

\item \code{lock_data}: Lock specified instruments

\item \code{unlock_data}: Unlock specified instruments
}}

