% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_key.R
\name{get_private_key}
\alias{get_private_key}
\alias{get_public_key}
\alias{get_private_key_path}
\alias{get_public_key_path}
\title{Get a private/public key of a package}
\usage{
get_private_key(package = gutils:::get_package_name(), devtools_load = FALSE)

get_public_key(package = gutils:::get_package_name(), devtools_load = FALSE)

get_private_key_path(
  package = gutils:::get_package_name(),
  devtools_load = FALSE
)

get_public_key_path(
  package = gutils:::get_package_name(),
  devtools_load = FALSE
)
}
\arguments{
\item{package}{(optional) a string indicating the target package. If not
assigned, the function will try to use the name of the active project
directory.}

\item{devtools_load}{(optional) a \code{\link[base:logical]{logical}} value
indicating if the function must call \code{devtools::load_all(".")} before
running. This is useful if you use \href{https://devtools.r-lib.org/}{devtools}
for package development. If \code{devtools_load = FALSE} and the dev package is
not loaded, the function will use a installed version of the package
(default: \code{FALSE}).}
}
\value{
A \code{\link[openssl:read_key]{rsa}} object with the data of a private/public
key or a string with the path of a private/public key.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{get_private_key()} and \code{get_public_key()} returns the a
\code{\link[openssl:read_key]{rsa}} object with the data of a private/public key
inside an R package.

\code{get_private_key_path()} and \code{get_public_key_path()} returns a string with
the location of a private/public key inside an R package.
}
\details{
The keys must be in a directory named \code{ssh} that must be present in the
package system folder (for \href{https://bit.ly/3F4Wqwz}{source packages} that is
usually the \code{inst} directory). The private and public keys must be named
\code{id_rsa} and \code{id_rsa.pub}, respectively.
}
\examples{
\dontrun{
get_private_key()
get_public_key()
get_private_key_path()
get_public_key_path()
}
}
\concept{utility functions}
