% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock_file.R
\name{lock_file}
\alias{lock_file}
\alias{unlock_file}
\title{Encrypt or decrypt single files}
\usage{
lock_file(
  file,
  public_key = here::here("_ssh", "id_rsa.pub"),
  suffix = ".lockr",
  remove_file = TRUE
)

unlock_file(
  file,
  private_key = here::here("_ssh", "id_rsa"),
  suffix = ".lockr",
  remove_file = TRUE,
  password = NULL
)
}
\arguments{
\item{file}{A \code{\link[base:character]{character}} string with the file path to be
encrypted/decrypted. For security reasons, encrypted files must end with
the suffix parameter.}

\item{public_key}{(optional) An \code{\link[openssl:keygen]{openssl}} RSA public
key or a \code{\link[base:character]{character}} string specifying the public key
path. See \code{\link[=rsa_keygen]{rsa_keygen()}} to learn how to create an RSA key pair
(default: \code{here::here("_ssh", "id_rsa.pub")}).}

\item{suffix}{(optional) A \code{\link[base:character]{character}} string specifying
the suffix to add when encrypting or remove when decrypting. Must start
with \code{.} (default: \code{".lockr"}).}

\item{remove_file}{(optional) A \code{\link[base:logical]{logical}} value indicating
if the original file must be removed after the encryption/decryption
process (default: \code{TRUE}).}

\item{private_key}{(optional) An \code{\link[openssl:keygen]{openssl}} RSA
private key or a \code{\link[base:character]{character}} string specifying the
private key path. See \code{\link[=rsa_keygen]{rsa_keygen()}} to learn how to create an RSA key pair
(default: \code{here::here("_ssh", "id_rsa")}).}

\item{password}{(optional) A \code{\link[base:character]{character}} string
specifying the password for decrypting a protected private key.
For security, avoid hardcoding passwords in scripts. Use
\code{\link[askpass:askpass]{askpass()}} to prompt for secure input
(default: \code{NULL}).}
}
\value{
An \link[base:invisible]{invisible} string containing the
locked/unlocked file path.
}
\description{
\code{lock_file()} and \code{unlock_file} encrypt or decrypt any kind of file using an
\href{https://www.openssl.org/}{OpenSSL} RSA key pair.
}
\examples{
## Creating Test Files and Keys -----

temp_dir <- tempfile("dir")
dir.create(temp_dir)

temp_file <- tempfile(tmpdir = temp_dir)

rsa_keygen(temp_dir)

con <- file(temp_file, "w+")
cat("Test", file = temp_file, sep = "\n")
list.files(temp_dir)
suppressWarnings(readLines(con))
close(con)

## Locking Files -----

temp_file |>
  lock_file(
    public_key = file.path(temp_dir, "id_rsa.pub")
  )

temp_file_locked <- paste0(temp_file, ".lockr")

con <- file(temp_file_locked, "rb")
list.files(temp_dir)
suppressWarnings(readLines(con))
close(con)

## Unlocking Files -----

temp_file_locked |>
  unlock_file(
    private_key = file.path(temp_dir, "id_rsa")
  )

list.files(temp_dir)
con <- file(temp_file, "r+")
readLines(con)
close(con)
}
\seealso{
Other lock/unlock functions: 
\code{\link{lock_dir}()},
\code{\link{save_and_lock}()}
}
\concept{lock/unlock functions}
