% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockr-package.R
\docType{package}
\name{lockr-package}
\alias{lockr}
\alias{lockr-package}
\title{lockr: Easily Encrypt/Decrypt Files}
\description{
A user-friendly toolkit to encrypt/decrypt files, specially made for research compendiums.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danielvartan.github.io/lockr/}
  \item \url{https://github.com/danielvartan/lockr/}
  \item Report bugs at \url{https://github.com/danielvartan/lockr/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

}
\keyword{internal}
