% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_level.R
\name{record_action}
\alias{record_action}
\alias{record_status}
\alias{lock_record}
\alias{unlock_record}
\title{Perform record-level locking operations}
\usage{
record_action(
  record,
  action = c("status", "lock", "unlock"),
  redcap_uri,
  token
)

record_status(record, redcap_uri, token)

lock_record(record, redcap_uri, token)

unlock_record(record, redcap_uri, token)
}
\arguments{
\item{record}{A vector of REDCap record names to affect. Required.}

\item{action}{A character string for the action to perform on records. One of "lock", "unlock" or "status"}

\item{redcap_uri}{A character string of the REDCap server's API endpoint URL}

\item{token}{A character string of the REDCap API token}
}
\value{
A dataframe with one row per record affected by the operation.
}
\description{
Perform record-level locking operations
}
\section{Functions}{
\itemize{
\item \code{record_status}: Return the lock status of record/s

\item \code{lock_record}: Lock named record/s

\item \code{unlock_record}: Unlock named record/s
}}

