% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsa_keygen.R
\name{rsa_keygen}
\alias{rsa_keygen}
\title{Generate and write a pair of RSA keys}
\usage{
rsa_keygen(dir = "./inst/ssh", password = NULL, bits = 2048)
}
\arguments{
\item{dir}{(optional) a string specifying the directory to save the generated
keys (default: \code{"./inst/ssh"})}

\item{password}{(optional) a string specifying the password to protect the
private key. If \code{NULL}, the private key will not be protected (default:
\code{NULL}).}

\item{bits}{(optional) an integer number specifying the length of the RSA key
in bits (default: \code{2048}).}
}
\value{
An invisible \code{NULL}. This function is called  for side effects only.
}
\description{
\code{rsa_keygen()} generates a pair of RSA (Rivest-Shamir-Adleman) private and
public keys for your project.
}
\examples{
temp_dir = tempfile("dir")
dir.create(temp_dir)

rsa_keygen(dir = temp_dir, password = "test", bits = 2048)

list.files(temp_dir)
}
\concept{utility functions}
