% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_lock.R
\name{save_and_lock}
\alias{save_and_lock}
\title{Save and lock a file}
\usage{
save_and_lock(
  x,
  file,
  type = "rds",
  public_key = here::here(".ssh", "id_rsa.pub"),
  ...
)
}
\arguments{
\item{x}{An R object to be saved and locked. If \code{type = "csv"}, \code{x} must be
a \code{\link[base:data.frame]{data.frame}}.}

\item{file}{A \code{\link[base:character]{character}} string indicating the file path
to save the object to.}

\item{type}{A \code{\link[base:character]{character}} string indicating the file type
to save the object as. Must be one of \code{"csv"} or \code{"rds"}
(default: \code{"rds"}).}

\item{public_key}{(optional) An \code{\link[openssl:keygen]{openssl}} RSA public
key or a \code{\link[base:character]{character}} string specifying the public key
path. See \code{\link[=rsa_keygen]{rsa_keygen()}} to learn how to create an RSA key pair
(default: \code{here::here(".ssh", "id_rsa.pub")}).}

\item{...}{(optional) Additional arguments passed to
\code{\link[readr:write_delim]{write_csv}} or \code{\link[readr:read_rds]{write_rds}}.}
}
\value{
An \link[base:invisible]{invisible} string containing the locked file
path.
}
\description{
\code{save_and_lock()} saves a data frame or R object to a file and then locks
it using \href{https://en.wikipedia.org/wiki/RSA_cryptosystem}{RSA} encryption.
}
\examples{
## Creating Keys and Test Object -----

temp_dir <- tempfile("dir")
dir.create(temp_dir)

rsa_keygen(temp_dir)

x <- letters

## Locking and Saving Objects -----

x |>
  save_and_lock(
    file = tempfile(),
    type = "rds",
    public_key = file.path(temp_dir, "id_rsa.pub")
  )
}
\seealso{
Other lock/unlock functions: 
\code{\link{lock_dir}()},
\code{\link{lock_file}()}
}
\concept{lock/unlock functions}
