% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticrypt.R
\name{staticrypt}
\alias{staticrypt}
\title{Encrypt or Decrypt HTML Files with StatiCrypt}
\usage{
staticrypt(
  file,
  password,
  config = NULL,
  directory = file.path(dirname(file), "encrypted"),
  decrypt = FALSE,
  recursive = FALSE,
  remember = 0,
  salt = NULL,
  share = NULL,
  share_remember = FALSE,
  template = NULL,
  template_button = "DECRYPT",
  template_color_primary = "#4CAF50",
  template_color_secondary = "#76B852",
  template_instructions = "",
  template_error = "Bad password!",
  template_placeholder = "Password",
  template_remember = "Remember me",
  template_title = "Protected Page",
  template_toggle_hide = "Hide password",
  template_toggle_show = "Show password"
)
}
\arguments{
\item{file}{A \code{\link[base:character]{character}} string indicating the path to
the HTML file to be encrypted or decrypted.}

\item{password}{A \code{\link[base:character]{character}} string specifying the
password to use for encryption or decryption. For security, avoid
hardcoding passwords in scripts. Use \code{\link[askpass:askpass]{askpass()}} to
prompt for secure input.}

\item{config}{A \code{\link[base:character]{character}} string specifying the path to
the configuration file. If \code{NULL}, defaults to a \code{.staticrypt.json}
file in the working directory (default: \code{NULL}).}

\item{directory}{A \code{\link[base:character]{character}} string indicating the
output directory for the encrypted or decrypted file (default: a subdirectory
named \code{encrypted} in the same directory as \code{file}). Use \code{dirname(file)}
to overwrite the file in place.}

\item{decrypt}{A \code{\link[base:logical]{logical}} flag indicating whether to
decrypt or encrypt the file (default: \code{FALSE}).}

\item{recursive}{A \code{\link[base:logical]{logical}} flag indicating whether to
process files in subdirectories recursively (default: \code{FALSE}).}

\item{remember}{A integer number specifying the number of days to remember
the password (default: \code{0}, meaning do not remember). Use \code{FALSE} to hide
the "Remember me" option in the generated HTML.}

\item{salt}{A \code{\link[base:character]{character}} string specifying a
32-character hexadecimal salt value for encryption. If \code{NULL}, a random
salt will be generated (default: \code{NULL}).}

\item{share}{A \code{\link[base:character]{character}} string specifying a path to
share the password securely (default: \code{NULL}, meaning do not share).}

\item{share_remember}{A \code{\link[base:logical]{logical}} flag indicating whether
to remember the shared password (default: \code{FALSE}).}

\item{template}{A \code{\link[base:character]{character}} string specifying the path
to a custom HTML template file (default: \code{NULL}, meaning use the default
template).}

\item{template_button}{A \code{\link[base:character]{character}} string specifying the
text for the submit button in the template (default: \code{"DECRYPT"}).}

\item{template_color_primary}{A \code{\link[base:character]{character}} string
specifying the template primary color in hexadecimal format
(default: \code{"#4CAF50"}).}

\item{template_color_secondary}{A \code{\link[base:character]{character}} string
specifying the template secondary color in hexadecimal format
(default: \code{"#76B852"}).}

\item{template_instructions}{A \code{\link[base:character]{character}} string
specifying instructions to display in the template (default: \code{""}).}

\item{template_error}{A \code{\link[base:character]{character}} string specifying the
error message to display for incorrect passwords
(default: \code{"Bad password!"}).}

\item{template_placeholder}{A \code{\link[base:character]{character}} string
specifying the placeholder text for the password input field
(default: \code{"Password"}).}

\item{template_remember}{A \code{\link[base:character]{character}} string specifying
the label for the "Remember me" checkbox in the template
(default: \code{"Remember me"}).}

\item{template_title}{A \code{\link[base:character]{character}} string specifying the
title of the protected page in the template (default: \code{"Protected Page"}).}

\item{template_toggle_hide}{A \code{\link[base:character]{character}} string
specifying the text for the "Hide password" toggle in the template
(default: \code{"Hide password"}).}

\item{template_toggle_show}{A \code{\link[base:character]{character}} string
specifying the text for the "Show password" toggle in the template
(default: \code{"Show password"}).}
}
\value{
An \link[base:invisible]{invisible} \code{NULL}. These functions are called
for side effects only.
}
\description{
\code{staticrypt()} offers an interface to the
\href{https://github.com/robinmoisson/staticrypt}{StatiCrypt} tool for encrypting
or decrypting HTML files using a password.
}
\examples{
\dontrun{
  library(askpass)

  staticrypt(
    file = "path/to/file.html",
    password = askpass()
  )
}
}
\concept{StatiCrypt functions}
