% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{encrypt_extdata}
\alias{encrypt_extdata}
\alias{decrypt_extdata}
\title{Encrypt/decrypt files from the \code{extdata} folder}
\usage{
encrypt_extdata(
  type = NULL,
  file = NULL,
  remove_file = TRUE,
  package = gutils:::get_package_name(),
  devtools_load = TRUE
)

decrypt_extdata(
  type = NULL,
  file = NULL,
  remove_file = TRUE,
  package = gutils:::get_package_name(),
  devtools_load = TRUE
)
}
\arguments{
\item{type}{(optional) a string indicating the folder of the \code{file} argument.
If \code{NULL} the function will encrypt/decrypt all files in the \code{extdata}
folder (default: \code{NULL}).}

\item{file}{(optional) a string indicating the file to encrypt/decrypt. If
\code{file} is \code{NULL} and \code{type} is assigned, the function will encrypt/decrypt
all files in the \code{type} folder (default: \code{NULL}).}

\item{remove_file}{(optional) a \code{\link[=logical]{logical}} value indicating if
the original file must be removed after the encryption/decryption (default:
\code{TRUE}).}

\item{package}{(optional) a string indicating the target package. If not
assigned, the function will try to use the name of the active project
directory (requires the \code{rstudioapi} package).}

\item{devtools_load}{(optional) a \code{\link[=logical]{logical}} value indicating if
the function must call \code{devtools::load_all(".")} before running. This is
useful if you use \link[devtools:devtools]{devtools} for package
development. If \code{devtools_load = FALSE} and the dev package is not loaded,
the function will use a installed version of the package.}
}
\value{
An invisible \code{NULL}. The functions are called just for side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{encrypt_extdata()} and \code{decrypt_extdata()} encrypt/decrypt files from
the \code{extdata} folder of an R package.
}
\details{
When decrypting, a dialog window will open asking for the key password for
each file that must be decrypted. You can only run \code{decrypt_extdata()} in
\link[=interactive]{interactive} mode.

The keys must be in a folder named \code{ssh} that must be present in the system
folder (\code{inst}). The private and public keys must be named \code{id_rsa} and
\code{id_rsa.pub}, respectively.
}
\examples{
\dontrun{
encrypt_extdata()
decrypt_extdata()
}
}
\concept{encrypt/decrypt functions}
