% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryptrpak-package.R
\docType{package}
\name{encryptrpak-package}
\alias{encryptrpak}
\alias{encryptrpak-package}
\title{encryptrpak: An R Package to Help Encrypt/Decrypt Files of R Packages}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

{encryptrpak} is an extension for the {encryptr} package, adding new features to help encrypt/decrypt files of R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipso.github.io/encryptrpak/}
  \item \url{https://github.com/gipso/encryptrpak/}
  \item Report bugs at \url{https://github.com/gipso/encryptrpak/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Interdisciplinary Sleep Research Group (GIPSO) [funder]
  \item University of Sao Paulo (USP) [funder]
}

}
\keyword{internal}
