% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryptrpak-package.R
\docType{package}
\name{encryptrpak-package}
\alias{encryptrpak}
\alias{encryptrpak-package}
\title{encryptrpak: Encrypt/Decrypt Files of R Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An extension for the 'encryptr' package, adding new features to encrypt/decrypt files of R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://giperbio.github.io/encryptrpak/}
  \item \url{https://github.com/giperbio/encryptrpak/}
  \item Report bugs at \url{https://github.com/giperbio/encryptrpak/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

}
\keyword{internal}
