% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_key.R
\name{get_private_key}
\alias{get_private_key}
\alias{get_public_key}
\title{Get a private/public key of a package}
\usage{
get_private_key(package = gutils:::get_package_name())

get_public_key(package = gutils:::get_package_name())
}
\arguments{
\item{package}{(optional) a string indicating the target package. If not
assigned, the function will try to use the name of the active project
directory (requires the \code{rstudioapi} package).}
}
\value{
The data of a private/public key.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{get_private_key()} and \code{get_public_key()} returns the data of a
private/public key inside an R package.
}
\details{
The keys must be in a folder named \code{ssh} that must be present in the system
folder (\code{inst}). The private and public keys must be named \code{id_rsa} and
\code{id_rsa.pub}, respectively.
}
\examples{
\dontrun{
get_private_key()
get_public_key()
}
}
\concept{utility functions}
