% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock_dir.R
\name{lock_dir}
\alias{lock_dir}
\alias{unlock_dir}
\title{Encrypt or decrypt all files in a directory}
\usage{
lock_dir(
  dir = "./inst/extdata",
  public_key = "./inst/ssh/id_rsa.pub",
  suffix = ".lockr",
  remove_file = TRUE
)

unlock_dir(
  dir = "./inst/extdata",
  private_key = "./inst/ssh/id_rsa",
  suffix = ".lockr",
  remove_file = TRUE,
  password = NULL
)
}
\arguments{
\item{dir}{A string indicating the directory to encrypt/decrypt. (default:
\code{"./inst/extdata"}).}

\item{public_key}{(optional) an \code{\link[openssl:keygen]{openssl}} RSA
public key or a string specifying the public key path. See \code{\link[=rsa_keygen]{rsa_keygen()}}
to learn how to create an RSA key pair (default:
\code{"./inst/ssh/id_rsa.pub"}).}

\item{suffix}{(optional) a string specifying the suffix to be added (when
encrypting)/removed (when decrypting) to/of the file. It must start with
\code{.} (default: \code{".lockr"}).}

\item{remove_file}{(optional) a \code{\link[=logical]{logical}} value indicating if
the original file must be removed after the encryption/decryption process
(default: \code{TRUE}).}

\item{private_key}{(optional) an \code{\link[openssl:keygen]{openssl}} RSA
private key or a string specifying the private key path. See \code{\link[=rsa_keygen]{rsa_keygen()}}
to learn how to create an RSA key pair (default: \code{"./inst/ssh/id_rsa"}).}

\item{password}{(optional) only for protected keys. A string specifying
the password to read the private key. Avoid typing passwords on the
console, use \link[askpass:askpass]{askpass()} instead (default: \code{NULL}).}
}
\value{
An invisible \code{NULL}. These functions are called just for side
effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\strong{WARNING}: This function must be used with caution! Check the parameters
carefully.

\code{lock_dir()} and \code{unlock_dir()} can encrypt or decrypt all files from a given
directory.
}
\examples{
## Locking files

ssh_dir <- tempfile("ssh")
dir.create(ssh_dir)
rsa_keygen(ssh_dir)

temp_dir <- tempfile("dir")
dir.create(temp_dir)
for (i in seq(5)) {
    file.create(tempfile(tmpdir = temp_dir))
}
list.files(temp_dir)

lock_dir(temp_dir, public_key = file.path(ssh_dir, "id_rsa.pub"))

## Unlocking files

unlock_dir(temp_dir, private_key = file.path(ssh_dir, "id_rsa"))
}
\seealso{
Other lock/unlock functions: 
\code{\link{lock_file}()}
}
\concept{lock/unlock functions}
